[WIP] Add more multiprocessing support
All demo files should now correctly set the number of CPUs to be used.
MKL and BLAS flags are placed at the beginning of every problem file as
a workaround to avoid over-subscription of cores.

In the future, we should move to using threadpoolctl.py (need to add as
a KADAL dependency) where we can more finely control which parts of the
code run with MKL/BLAS threads. (Mainly, where we manually parallelise
code with multiprocessing vs where we run a section in serial with
vectorised code.)

Other fixes:
 - Fix bug in akmcs.py where array was instantiated in reverse.
 - A bit more work on docstrings.